// ui/src/main/angular/ui/modules/lipidcompass-backend-client/model/cvParameterFacet.ts
/**
 * LipidCompass Backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CvParameterFacetValue } from './cvParameterFacetValue';


export interface CvParameterFacet { 
    referenceType?: string;
    facetValues?: Array<CvParameterFacetValue>;
}