// ui/src/main/angular/ui/modules/lipidcompass-backend-client/model/cvParameterFacetValue.ts
/**
 * LipidCompass Backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CvParameterFacetValue { 
    accession?: string;
    name?: string;
    count?: number;
}